Converted Modal component to typescript. Added built in button row and made the header, close button, and button row optional.
