+++ src/components/Modal/modal.scss
@@ -1,0 +1,7 @@
+@import "~vanilla-framework/scss/settings";
+
+.p-modal__button-row {
+  border-top: 1px solid $color-mid-light;
+  padding-top: 1rem;
+  text-align: right;
+}
